requirements.txt
Flask==2.2.2
functions-framework==3.*
gunicorn==20.1.0
python-telegram-bot==13.14
requests==2.28.1
youtube-dl==2021.12.17